Usando fragment e navigation para main activity
