+++ .idea/misc.xml
@@ -4,10 +4,11 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/drawable/image_placeholder.xml" value="0.134375" />
-        <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/activity_main.xml" value="0.2" />
+        <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/activity_main.xml" value="0.25" />
         <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/favorites_fragment.xml" value="0.3625" />
         <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/favorites_fragment2.xml" value="0.3625" />
         <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/films_fragment.xml" value="0.3625" />
+        <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/fragment_main.xml" value="0.3625" />
         <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/series_fragment.xml" value="0.3625" />
         <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/show_item.xml" value="0.5" />
       </map>
