+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:name=".BaseApplication"
         android:theme="@style/Theme.ZygoTV">
-        <activity android:name="com.zygotecnologia.zygotv.ui.main.MainActivity">
+        <activity android:name="com.zygotecnologia.zygotv.ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
