+++ app/src/main/java/com/zygotecnologia/zygotv/ui/MainActivity.kt
@@ -1,0 +1,18 @@
+package com.zygotecnologia.zygotv.ui
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.zygotecnologia.zygotv.databinding.ActivityMainBinding
+
+class MainActivity : AppCompatActivity() {
+
+    private lateinit var binding : ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        binding = ActivityMainBinding.inflate(layoutInflater)
+
+        setContentView(binding.root)
+    }
+}
