+++ app/src/main/res/layout/activity_main.xml
@@ -8,102 +8,18 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.main.MainActivity" >
+        tools:context=".ui.MainActivity" >
 
-        <TextView
-            android:id="@+id/tv_title_show_popular"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/spacing_24"
-            android:textColor="@color/white"
-            android:textSize="@dimen/text_size_16"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <com.google.android.material.card.MaterialCardView
-            android:id="@+id/cv_show_popular"
+        <androidx.fragment.app.FragmentContainerView
+            android:id="@+id/nav_host_fragment"
+            android:name="androidx.navigation.fragment.NavHostFragment"
             android:layout_width="match_parent"
-            android:layout_height="@dimen/item_list_poster_path_height"
-            android:layout_marginStart="@dimen/spacing_24"
-            android:layout_marginEnd="@dimen/spacing_24"
-            android:layout_marginTop="@dimen/spacing_8"
-            app:cardCornerRadius="@dimen/corner_radius_default"
+            android:layout_height="0dp"
+            app:defaultNavHost="true"
+            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/tv_title_show_popular" >
-
-            <ImageView
-                android:id="@+id/iv_show_popular"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:scaleType="center"
-                android:src="@drawable/image_placeholder" />
-
-        </com.google.android.material.card.MaterialCardView>
-
-        <TextView
-            android:id="@+id/tv_title_crime"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/spacing_24"
-            android:layout_marginTop="@dimen/spacing_24"
-            android:text="@string/category_title_crime"
-            android:textColor="@color/white"
-            android:textSize="@dimen/text_size_16"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/cv_show_popular" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rv_show_list_crime"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="@dimen/spacing_24"
-            android:orientation="horizontal"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintTop_toBottomOf="@+id/tv_title_crime"
-            tools:listitem="@layout/show_item" />
-
-        <TextView
-            android:id="@+id/tv_title_action_adventure"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/spacing_24"
-            android:text="@string/category_title_action_and_adventure"
-            android:textColor="@color/white"
-            android:textSize="@dimen/text_size_16"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/rv_show_list_crime" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rv_show_list_adventure"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="@dimen/spacing_24"
-            android:orientation="horizontal"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintTop_toBottomOf="@id/tv_title_action_adventure"
-            tools:listitem="@layout/show_item" />
-
-        <TextView
-            android:id="@+id/tv_title_comedy"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/spacing_24"
-            android:text="@string/category_title_comedy"
-            android:textColor="@color/white"
-            android:textSize="@dimen/text_size_16"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/rv_show_list_adventure" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rv_show_list_comedy"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="@dimen/spacing_24"
-            android:orientation="horizontal"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintTop_toBottomOf="@id/tv_title_comedy"
-            tools:listitem="@layout/show_item" />
+            app:navGraph="@navigation/navigation_main" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
